updated routing, bounties page, header, footer etc
