+++ www/index.html
@@ -3,7 +3,7 @@
     <head>
         <title>Ritocoin</title>
         <meta name="viewport" content="width=device-width, initial-scale=1">
-        <link rel="stylesheet" href="./build/ritocoin.css?s=3"></head>
+        <link rel="stylesheet" href="./build/ritocoin.css"></head>
     <body>
         <noscript>
         You need to enable JavaScript to run this app.
@@ -11,6 +11,6 @@
 
         <div id="root"></div>
 
-        <script src="./build/ritocoin.js?s=3"></script>
+        <script src="./build/ritocoin.js"></script>
     </body>
 </html>
