+++ www/src/components/App.jsx
@@ -1,9 +1,6 @@
 import React from 'react';
-import { Router, Route, Switch } from 'react-router-dom';
+import { HashRouter, Route, Switch } from 'react-router-dom';
 
-import history from '../utils/history';
-
-import Layout from './layout/Layout';
 import Home from './Home';
 import Bounties from './Bounties';
 
@@ -11,13 +8,12 @@ import '../less/main.less';
 
 export default function App() {
     return (
-        <Router history={history}>
-            <Layout>
-                <Switch>
-                    <Route exact path="/" component={Home} />
-                    <Route exact path="/bounties" component={Bounties} />
-                </Switch>
-            </Layout>
-        </Router>
+        <HashRouter>
+            <Switch>
+                <Route exact path="/" component={Home} />
+                <Route exact path="/bounties" component={Bounties} />
+                <Route path="*" component={Home} />
+            </Switch>
+        </HashRouter>
     );
 }
