+++ www/src/less/modules/bounties.less
@@ -51,6 +51,10 @@
                     padding: 5px 2%;
                 }
             }
+            @media only screen and (max-width: 500px) {
+                width: 32%;
+                font-size: 3vw;
+            }
         }
 
         @media only screen and (max-width: 1100px) {
@@ -64,16 +68,20 @@
             overflow: hidden;
             font-size: 14px;
             margin-bottom: 14px;
+            
+            @media only screen and (max-width: 500px) {
+                font-size: 3vw;
+            }
 
             .item {
                 float: left;
                 border-left: 1px solid @ritoGray;
-                padding: 5px 2%;
-                width: 15%;
+                padding: 5px 1%;
+                width: 17.8%;
                 color: @ritoGray;
                 overflow: scroll;
                 cursor: ew-resize;
-                white-space: nowrap;
+                word-break: break-word;
 
                 @media only screen and (max-width: 1100px) {
                     float: none;
