+++ www/src/less/modules/footer.less
@@ -7,6 +7,10 @@ footer {
     width: 100%;
     height:80px;
 
+    @media only screen and (max-width: 600px) {
+        position: relative;
+    }
+
     p {
         font-size: 14px;
         line-height: 1.5em;
