+++ www/src/less/modules/header.less
@@ -5,6 +5,37 @@ header {
     overflow: hidden;
     position: relative;
 
+    .links {
+        height: 30px;
+        background-color: @lightestBlue;
+
+        .more-info {
+            position: absolute;
+            top: 5px;
+            right: 60px;
+    
+            @media only screen and (max-width: 500px) {
+                margin-left: auto;
+                margin-right: auto;
+                left: 0;
+                right: 0;
+                width: 210px;
+            }
+    
+            a {
+                cursor: pointer;
+                padding: 0 20px;
+    
+                svg {
+                    width:20px;
+                    fill: @white;
+                }
+    
+            }
+    
+        }
+    }
+
     .page-title {
         float: left;
 
@@ -21,30 +52,25 @@ header {
             color: @white;
         }
 
-    }
+        ul {
+            overflow: hidden;
+            position: absolute;
+            right: 60px;
+            bottom: 0;
+            
+            li {
+                float: left;
+                list-style: none;
+                padding: 0 10px;
+                font-size: 20px;
+                font-weight: 300;
+                cursor: pointer;
 
-    .more-info {
-        position: absolute;
-        bottom: 10px;
-        right: 10px;
-
-        @media only screen and (max-width: 500px) {
-            margin-left: auto;
-            margin-right: auto;
-            left: 0;
-            right: 0;
-            width: 210px;
-        }
-
-        a {
-            cursor: pointer;
-            padding: 0 20px;
-
-            svg {
-                width:30px;
-                fill: @white;
+                a {
+                    color:@white;
+                    text-decoration: none;
+                }
             }
-
         }
 
     }
