+++ www/src/utils/bountyPayouts.js
@@ -7,11 +7,11 @@ const bounties = [
         txId: '30dedef29576e31521966a4deb622561efcd370a6863ad4c5105fc3343766a81',
     },
     {
-        who: 'Sir_Chasington (discord user)',
-        what: 'Website initial framework update',
-        amount: '5000',
-        receivingAddress: 'B9q1Mno66GCuTejSz9UHdVWqzdZFnFp9RV',
-        txId: '30dedef29576e31521966a4deb622561efcd370a6863ad4c5105fc3343766a81',
+        who: 'Under (discord user)',
+        what: 'Launch Work. Helped track down compile problems, provided complete workflow for builds for all architectures.',
+        amount: '100,000',
+        receivingAddress: 'B6XGRhcGUo8dY1Chh5VGCzceNBNxDAtSPP',
+        txId: 'da5bce9f9c6afba0039860788fda4430468dafe788cf6632154a1a321f80bfe5',
     },
 ];
